### 2024-08-14 #249 @ethanwhite
A few notes on the editorialbot's checks:

1. There is a Statement of Need section, it's just titled Introduction and Statement of Need since the section covers both. Let us know if we need to figure out how to separate them.
2. Most of the missing DOI's are for things that don't have DOI's. The CRAN citations do have them now and we'll add them during revision.

Suggested reviewers:

* Sarah Masud (sara-02) - Probably the best reviewer on the JOSE list due to interest in offline tools and experience with Python
* John Ray (johnlray) - Also on the JOSE list and has an interest in web scraping (which this package does a lot of) and experience with Python
* Angelique Trusler (elletjies) Not on the JOSE list, but a good reviewer given experience organizing training in internet-limited environments

Some other folks on the JOSE list with experience in Python and teaching:
* Kristian Rother (krother)
* Ariel Roke (arokem)

Let us know if there's anything else we can do or provide to help with the review process.
