### 2024-09-03 2f136df
Update version for patch release to fix wheel build
See #140
