### 2024-10-01 0419225
Initial work on custom lessons
Currently works, but index page fails to link if custom lesson root is a
2nd level (or greater) subdirectory.
Likely also fails if it is not in a 1st level subdirectory.
