+++ offlinedatasci/main.py
@@ -92,8 +92,8 @@ def download_and_save_installer(latest_version_url, destination_path):
     destination_path -- Path to save installer
     """
     if not os.path.exists(destination_path):
-                print("****Downloading file: ", destination_path)    
-                urllib.request.urlretrieve(latest_version_url, destination_path) 
+        print("****Downloading file: ", destination_path)
+        urllib.request.urlretrieve(latest_version_url, destination_path)
     else:
         print("File not being downloaded")
 
@@ -114,7 +114,7 @@ def download_r(ods_dir):
     download_r_macosx(r_current_version, ods_dir)
 
 
-def download_lessons(ods_dir):
+def download_lessons(ods_dir, lessons = "all"):
     """Downloads the workshop lessons as rendered HTML.
     Keyword arguments:
     destination_path -- Path to save rendered HTML lessons
@@ -132,56 +132,58 @@ def download_lessons(ods_dir):
         """)
         return
 
-    dc_lessons = ["https://datacarpentry.org/ecology-workshop/",
-                  "https://datacarpentry.org/spreadsheet-ecology-lesson/",
-                  "http://datacarpentry.org/OpenRefine-ecology-lesson/",
-                  "https://datacarpentry.org/R-ecology-lesson/",
-                  "https://datacarpentry.org/python-ecology-lesson/",
-                  "https://datacarpentry.org/sql-ecology-lesson/"]
-    lc_lessons = ["https://librarycarpentry.org/lc-overview/",
-                  "https://librarycarpentry.org/lc-data-intro/",
-                  "https://librarycarpentry.org/lc-shell/",
-                  "https://librarycarpentry.org/lc-open-refine/",
-                  "https://librarycarpentry.org/lc-git/",
-                  ]
+    dc_lessons = {
+        "https://datacarpentry.org/ecology-workshop/": "data-carpentry",
+        "https://datacarpentry.org/spreadsheet-ecology-lesson/": "data-carpentry",
+        "http://datacarpentry.org/OpenRefine-ecology-lesson/": "data-carpentry",
+        "https://datacarpentry.org/R-ecology-lesson/": "data-carpentry",
+        "https://datacarpentry.org/python-ecology-lesson/": "data-carpentry",
+        "https://datacarpentry.org/sql-ecology-lesson/": "data-carpentry"
+    }
+    lc_lessons = {
+        "https://librarycarpentry.org/lc-overview/": "library-carpentry",
+        "https://librarycarpentry.org/lc-data-intro/": "library-carpentry",
+        "https://librarycarpentry.org/lc-shell/": "library-carpentry",
+        "https://librarycarpentry.org/lc-open-refine/": "library-carpentry",
+        "https://librarycarpentry.org/lc-git/": "library-carpentry"
+    }
+    sc_lessons = {
+        "http://swcarpentry.github.io/shell-novice": "software-carpentry",
+        "http://swcarpentry.github.io/git-novice": "software-carpentry",
+        "http://swcarpentry.github.io/python-novice-inflammation": "software-carpentry",
+        "http://swcarpentry.github.io/python-novice-gapminder": "software-carpentry",
+        "http://swcarpentry.github.io/r-novice-inflammation": "software-carpentry",
+        "http://swcarpentry.github.io/r-novice-gapminder": "software-carpentry",
+        "http://swcarpentry.github.io/shell-novice-es": "software-carpentry",
+        "http://swcarpentry.github.io/git-novice-es": "software-carpentry",
+        "http://swcarpentry.github.io/r-novice-gapminder-es": "software-carpentry"
+    }
+
+    lessons_to_install = {}
+    for lesson in lessons:
+        if lesson == "all":
[…]
