+++ offlinedatasci/__init__.py
@@ -1,4 +1,4 @@
 from . import main
 from .main import *
 
-__version__ = '0.6.1'
+__version__ = '0.7.0'
