+++ pyproject.toml
@@ -6,7 +6,7 @@ build-backend = "setuptools.build_meta"
 name = "offlinedatasci"
 description = "Download and configure common tools for teaching and doing data science without an internet connection"
 readme = "README.md"
-version = "0.6.1"
+version = "0.7.0"
 
 authors = [
   { name="Virnaliz Cruz" },
