### 2025-08-11 #249 @nicoguaro
@sara-02, @johnlray, would you review this submission for JOSE? It is a set of _tools for teaching and doing data science without an internet connection_.
