### 2025-09-05 1ab3a8a
Add testing dependencies including those for new activate functionality
