### 2025-10-30 #249 @nicoguaro
@elletjies, would you review this submission for JOSE? It is a _set of tools for teaching and doing data science without an internet connection_.
