### 2025-11-04 #249 @nicoguaro
@elletjies, your work within The Carpentries might prove really useful here.

The purpose of reviews is to help authors improve the quality of the submission to the standard for acceptance into JOSE (from [docs](https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html)). Regarding the criteria, we have a checklist with the following categories

- Conflict of interest
- Code of Conduct
- General checks
- Documentation
- Pedagogy / Instructional design (Work-in-progress: reviewers, please comment!)
- JOSE paper

You can see [an example here](https://github.com/openjournals/jose-reviews/issues/293#issuecomment-3238898564).

We do not have a hard due date, but I would suggest as a goal the first week of December.

Does this work for you?

Do not hesitate to ping me if you have more questions (or contact me via email).
