### 2025-11-17 #249 @ethanwhite
(co-author) I agree that @elletjies would have a lot to contribute in reviewing this package.

Another good option would be jcohen02 https://profiles.imperial.ac.uk/jeremy.cohen
