### 2025-11-18 #249 @nicoguaro
@jsteyn, would you like to review this submission for JOSE? It is a _set of tools for teaching and doing data science without an internet connection_.
