### 2025-11-18 #249 @ethanwhite
@jsteyn is an author of the paper (definitely well qualified!)
