### 2025-11-18 #249 @nicoguaro
> [@jsteyn](https://github.com/jsteyn) is an author of the paper (definitely well qualified!)

I see. I didn't check. I will send Jeremy Cohen an email inviting him.
