### 2026-01-06 #249 @jcohen02
Hi all, thanks for the invitation from @nicoguaro to review this submission. I’m happy to do this but for transparency, and as I mentioned in our email exchange, I note that I do know some of the authors of this submission (as members of the Research Software Engineering community) and plan to collaborate with a couple of them on an event in this field in the coming months. It's my understanding that this doesn't present an issue so I will go ahead and review the manuscript.
