### 2026-01-07 #322 @editorialbot
[REVIEW]: offlinedatasci: A Python Package for Managing Data Science Software Installers when Limited Access to the Internet is Anticipated
**Submitting author:** @garezana (<a href="http://orcid.org/0000-0001-6783-2651">Virnaliz Cruz</a>)
**Repository:** https://github.com/carpentriesoffline/offlinedatasci/
**Branch with paper.md** (empty if default branch):
**Version:** v0.7.0
**Paper kind:** software
