### 2026-01-07 #322 @nicoguaro
@sara-02, @jcohen02, this is the space where the review process takes form. There is a checklist for each one, tick the boxes when you see that the criterion is satisfied. You can generate the checklist with

I will be here to answer the questions that you might have.

Let us use as a tentative timeframe the last week of January, is that OK for you?
