### 2026-01-07 #322 @jcohen02
##### Initial comments

This submission represents an interesting software package to support data-science-related training in environments with limited or no internet connectivity. The package provides functionality to download a set of software dependencies and Carpentries training material from a range of different sources, including setting up partial R/Python package repository mirrors, to enable offline use of the training materials, with access to all necessary software tools and dependencies. The `offlinedatasci` tool also supports incremental updating of downloaded materials.

The provided paper is well written and structured and contains extensive detail about the tool. The tool appears to function successfully as detailed in the paper/repository. It's clear that this tool supports an important capability in taking data science-related training opportunities to a wider and more diverse audience.

Beyond the paper and software package, there are a number of areas where improvements could be made, both to address the requirements of the review checklist below and to enhance the information available to users of the package. Detailed comments have been provided between the checklist and a number of issues created in the software's GitHub repository. Some further summary comments are provided below the checklist.

- [x] As the reviewer I confirm that I have read the [JOSE conflict of interest policy](https://github.com/openjournals/joss/blob/jose/COI.md) and that there are no conflicts of interest for me to review this work
    - _Please see [this comment](https://github.com/openjournals/jose-reviews/issues/249#issuecomment-3715826569) in the pre-review discussion. This has been checked with the editor who has confirmed that it's acceptable to proceed with this review._

    - An MIT licence is included in the repo. This is fine for the code. However, the paper is CC-BY and it may also make sense to licence the text-based docs under CC-BY too? See issue for further comments/suggestions:
       https://github.com/carpentriesoffline/offlinedatasci/issues/150

- [ ] **Version:** Does the release version given match the GitHub release?
    - The version referenced in this issue is v0.5.0 whereas the latest release in the repo seems to be v0.7.0. Not sure what, if anything, needs updating here since v0.5.0 is not mentioned anywhere in the paper itself, but the authors should, nonetheless, double check that the paper content still reflects the software's capabilities and functionality in light of repo changes since the original submission.

   - The list of contributors to the repository matches the author list. The majority of repository contributions have been made by two of the authors, one of whom is the submitting author. In this context, authorship seems appropriate. It is noted that at least 2 of the authors have made limited contributions to the repository in terms of commits, however it's assumed they've been engaged with and contributing to this work in different ways, e.g. they have been opening/contributing to issues which also show that they have been supporting and undertaking testing of the tool.
   - I also note that author @hturner seems to not be listed under authors in the package metadata (`pyproject.toml`). (Flagging this in case it's an oversight.)

_Testing was carried out under Linux with Ubuntu 22.04 with Python 3.10_

    - Installation from PyPi via `pip` and `pipx` works fine (tested using Python 3.10, since this is the most recent version listed in the package metadata).
      Installation of the latest code from GitHub as described in the two options under "_Installing development versions_" also works as expected.

    - This is largely OK but I have a couple of suggested updates that I think it would be good to address before this section is "signed off". Issues have been opened for these points and include further detail:
        - I'd suggest extending the information provided in the help text a little so that users are clear on what different options do and don't need to refer back to the documentation
           https://github.com/carpentriesoffline/offlinedatasci/issues/151
        - It would be helpful to provide more detailed output when running certain aspects of the `offlinedatasci install` task.
           https://github.com/carpentriesoffline/offlinedatasci/issues/152

    - There are no specific performance claims made.

#### Review feedback summary

Documentation is provided via readthedocs and, as noted in the paper, a workflow is in place to automate deployment of updated documentation when changes are made in the repository. The current documentation covers the core functionality of the library but based on the review points below, there are some things missing. In general, it feels like the documentation would benefit from some more descriptive text around the examples and some further detail of how the library can be used in a workflow for offline training as part of the wider carpentriesoffline infrastructure.

    - The readthedocs documentation does not currently include a statement of need. (This is clearly provided in the paper itself - maybe a link to the paper would suffice, although I think providing this information directly in the docs would be preferable).
       https://github.com/carpentriesoffline/offlinedatasci/issues/156

    - Some minor updates to the installation instructions are suggested:
       https://github.com/carpentriesoffline/offlinedatasci/issues/157

    - Example usage is provided in both the documentation and the repository README:
        - I feel that the the examples in the documentation would benefit from a little more detail.
           https://github.com/carpentriesoffline/offlinedatasci/issues/158 (this issue is also linked in the functionality section below)
        - A couple of the examples in the README also don't work and need updating (alongside some further README updates).
          https://github.com/carpentriesoffline/offlinedatasci/issues/155

    - Functionality is documented but I think this documentation would benefit from some further detail and context. There are some further enhancements that I don't think are vital in addressing this point but would contribute to further assisting users of the package. Details are provided in the issue created for this item:
       https://github.com/carpentriesoffline/offlinedatasci/issues/158

    - An automated test workflow is in place using GitHub Actions. This satisfies this aspect of the requirements but it would also be nice to have a way for developers contributing to the library to run a test suite locally and for this to be detailed in the docs/README - I'm not sure if this is possible with the current structure or how it would be done?

    - I can't see that this information is present in the documentation. An issue has been added to the repository:
       https://github.com/carpentriesoffline/offlinedatasci/issues/159

     - The paper provides a detailed and realistic statement of need. There is also an additional section covering software design which complements the statement of need.
    - The paper is well written and structured but it's significantly longer than the guideline length of around 1000 words. I'll leave it with the editors to comment on whether there's an issue here with length and the paper needs shortening significantly - from my perspective, I think the content provided is useful and helpful so I don't have an issue with the length in principle, however, I have a couple of suggested changes/updates that should result in slightly shortening the manuscript anyway. See repository issue for further details:
       https://github.com/carpentriesoffline/offlinedatasci/issues/153

    - There are no DOIs provided. This is largely because much of the referenced material does not appear to have DOIs available, however, some DOIs are missing as detailed in the relevant repository issue:
       https://github.com/carpentriesoffline/offlinedatasci/issues/154

 - [ ] A handful of other minor paper updates are suggested:
        https://github.com/carpentriesoffline/offlinedatasci/issues/160

#### Final review comments

As noted at the top of this review, the paper submitted via the `offlinedatasci` repository is well written and structured and contains extensive detail about the tool. A significant number of comments and suggestions have been provided above and these relate to text-based elements of the repository content - the README file and documentation.

There are lots of suggested updates but many of these should be straightforward and fairly quick to address. In this context, I'm unsure whether to grade this as minor or major revisions, but given that the paper itself requires minimal updates (largely fixing the references, removing the installation/API information and possibly optimising the text a little to reduce word count) and the software appears to work as described, I'm inclined towards recommending minor revisions. I also note that some of the content suggested for addition to the docs is already present in the paper itself so it's clear that the authors will be able to address these points within the documentation.

A couple of final points which are probably beyond the scope of the review but could be useful for the authors to consider in terms of future updates to the tool:

 - **Top-level index.html**: I think having a top-level `index.html` file generated alongside the downloaded content (a bit like what's done with the lesson material) would be very useful.

 - **Embedded server functionality**: I also think having some means of starting a small server via the `offlinedatasci` command to make it easy to access the downloaded content via a browser (and the aforementioned top-level index.html) would be a really useful enhancement. It may be that this can be achieved through the use of other carpentriesoffline tools, but if that's the case, it would be good to highlight it somewhere.

 - **Decoupling of data download locations from code**: I wonder if the authors have looked at/considered separating the many download URLs that are currently embedded within the code into a separate metadata/URLs file and then parameterising the download code somewhat. In this context, I think the design of the code could perhaps be enhanced to better support maintenance and extension. I don't think this is critical in the context of this submission so I've not opened an issue for this, but I think it would be an interesting and potentially beneficial enhancement to consider.
