### 2026-01-07 carpentriesoffline/offlinedatasci#150 @jcohen02
JOSE review - Licence for paper and documentation
The repository has an MIT licence applied. This is fine for the code but the paper that is also stored within this repository is, I believe, under a CC-BY licence and CC-BY or similar would be more appropriate for the documentation too.

Thoughts / suggestions on how to address this:

 - Add an additional licence file (perhaps something like `LICENSE.docs`?) that contains the CC-BY-4.0 licence and notes which elements of the repository this applies to (or this note could be provided via the following bullet point).
 - Add a "License" section to the end of the README that highlights that repository code is under an MIT licence and the paper and docs are CC-BY.
