### 2026-01-07 carpentriesoffline/offlinedatasci#151 @jcohen02
JOSE review - Update CLI help text
The `offlinedatasci` command help text covers all options but there's limited information about what the options actually do. A few suggestions of how this could be improved:

 - Add a short text string at the top stating what the command actually does (along the lines of help output provided by common Linux/Unix commands, e.g. see `ls --help`, `date --help`, etc.)
 - For the subcommands, what do _install_ and _add_ do (in the top-level help)
 - Within the subcommand help (both _install_ and _add_, various positional arguments are provided but it's not stated what each of these does. e.g. under `offlinedatasci install --help`:

    ```
    ...
    positional arguments:
       {all,lessons,r-packages,python,python-packages,r,rstudio}

    options:
       -h, --help            show this help message and exit
    ```
    maybe the positional arguments can be presented in a list similar to `-h` under options with a brief note for each?
