### 2026-01-07 carpentriesoffline/offlinedatasci#152 @jcohen02
Provide more detailed output for install / add commands
The `offlinedatasci install all` task works successfully. However, if possible, I think that providing more detailed output when running the `offlinedatasci install` task would be very helpful.

This issue is motivated by my experience when testing this element of the tool:

  - This would be useful especially during the part where the lesson download steps are carried out.
  - When running the installation, I found that there was vast quantities of output flying up the screen as all the R packages were built and installed and then it appeared to hang at "_Downloading lesson from https://datacarpentry.org/ecology-workshop/_". After several minutes I was convinced the install was stuck. I could see a `wget` process running but this didn't seem to be consuming CPU resource and I assumed that this had perhaps got blocked waiting to download something. When I looked again after several more minutes, the download had moved on to the next lesson but I was close to canceling the process!
  - A simple solution would possibly be to run `wget` with a verbose output switch. However, I guess this would generate large quantities of output, I wonder if there's a less verbose way of providing some feedback that something is happening when the lesson downloads are running?
