### 2026-01-08 carpentriesoffline/offlinedatasci#153 @jcohen02
JOSE review - paper updates
The JOSE paper within this repository is well written and structured but it's significantly longer than the [guideline length](https://openjournals.readthedocs.io/en/jose/submitting.html#what-should-my-paper-contain) of around 1000 words. I'll leave it with the editors to comment on whether there's an issue here with length and the paper needs shortening significantly - from my perspective, I think the content provided is useful and should be of interest to the reader so I don't have an issue with the length in principle. However, I have a couple of suggested changes/updates that should result in shortening the manuscript anyway:

- I think the text in the statement of need and (in particular) the software design sections could be cut back a little. Some of this content could be moved into the online documentation which is currently quite light on descriptive text - maybe a background section there could be used to cover some of this content?

- I note that the review guidance states "_...software paper’s should not include software documentation such as API (Application Programming Interface) functionality, as this should be outlined in the software documentation._" so I think the subsections on "Installation", "Command line interface" and "Python interface" should be removed from the paper and a note can simply be placed in the "Documentation" section advising readers to see the online documentation for installation details and command line interface and API details.
