### 2026-01-09 carpentriesoffline/offlinedatasci#154 @jcohen02
JOSE review - references
References in the JOSE paper included in this repository require updating to include DOIs in order to meet the requirements of the journal:

- [ ] The reference on line 226 is missing a DOI (10.7551/mitpress/11140.001.0001)

- [ ] The reference on lines 227-232 is also missing its DOI (10.3233/978-1-61499-649-1-87). There is also an incorrectly rendering hyphen in this reference.

- The references to R packages on lines 220-222 for the "remotes" R package and L247-L249 for the "miniCRAN" package: These package references have a note on their pages stating "_Please use the canonical form https://CRAN.R-project.org/package=[package name] to link to this page._", on this basis, if provided in the final version, these links should use the requested format. However, the [reference check](https://github.com/openjournals/jose-reviews/issues/322#issuecomment-3718795454) in the review thread has suggested DOIs for these items, and I'd recommend updating these 3 items to use the DOIs suggested from the reference check:
  - [ ] Add DOI for "remotes" R package (L220-L222)
  - [ ] Add DOI for "miniCRAN" R package (L247-L249)
  - [ ] Add DOI for R manuals (L237-L238) (if the suggested DOI is indeed pointing to the correct resource)
