### 2026-01-09 carpentriesoffline/offlinedatasci#155 @jcohen02
JOSE review - README updates
The README feels like it could benefit from a little more information:

#### Summary

The installation section of the README is basic (but clear). Most of the commands listed work, however, some don't and some updates are required - see below. I think there's also scope for a number of small but helpful updates to the README to provide a bit more information, including some additional background on what the package does and why it might be useful, and some detail on what the commands listed under "Installation" and "Usage" actually do.

#### Suggested updates / changes

- [ ] Fix the outdated command examples in the README that don't currently work (the main documentation contains correct versions of these commands) - `offlinedatasci add-packages`, `offlinedatasci install minicran`

- [ ] Add a couple of sentences for each of the examples under "Installation" and "Usage" to make it clear what the commands shown do.

- [ ] Add something that highlights the data volumes that will be downloaded when running the example commands. As a package that's designed to support operation in environments with no/limited internet connectivity, there may be cases where users need to make specific plans to be somewhere where there's sufficient bandwidth to download the materials, or consider things like data charges. This may not be immediately obvious to potential users from the current README/documentation content. e.g. after running the `install all` command, the target folder contains 2.2+GB of data (using Ubuntu Linux). This includes a range of generated artifacts from building native code elements of the libraries installed, however the paper suggests a 2023 figure of ~1.4GB data being downloaded and this may now be greater.

- [ ] Add something about supported Python versions for the library - the current package metadata suggests the package supports Python 3.6-3.10 but the test workflow currently appears to be running with 3.14.

- [ ] Consider adding a "License" section to the README (in addition to the LICENSE file(s) in the repo) that highlights the licence(s) that the code/docs/paper are under.

- [ ] Are there any funder acknowledgements to add? If so these could be included in an acknowledgements section of the README.
