### 2026-01-09 #322 @nicoguaro
@jcohen02, thanks for the quick and thorough review!

I will update the version of the software in the review issue.
