### 2026-02-13 carpentriesoffline/offlinedatasci#150 @jcohen02
Hi @ethanwhite, thanks for the explanation here. I highlighted this having seen prior examples of dual licensed repos (e.g. Carpentries lessons) and I thought this model would perhaps be most suitable here.
At the same time, if there's precedent/existing best practice that suggests an alternative would be more suitable, then there's no problem from my side if you want to take that route. I agree JOSS represents a good example here and if the approach taken there is to stick with a single licence that covers all the repo content, and then the published version of the paper is under a CC-BY licence, that seems reasonable.
If you want to stick with the approach you are already using of having a single licence on the repo, then feel free to close this issue as complete on the basis of this discussion. Thanks.
