### 2026-02-16 carpentriesoffline/offlinedatasci#150 @ethanwhite
Thanks @jcohen02! We'll go with the standard JOSS approach.
