### 2026-02-16 b47364c
Add missing DOIs for papers and R packages
Closes #154
