+++ paper.bib
@@ -9,7 +9,8 @@
   title = {Data Science},
   author = {Kelleher, John D. and Tierney, Brendan},
   year = {2018},
-  publisher = {MIT Press}
+  publisher = {MIT Press},
+  doi = {10.7551/mitpress/11140.001.0001}
 }
 
 @article{swenson2021internet,
@@ -25,7 +26,8 @@
   author = {Vries, Andrie de and Chubaty, Alex and Microsoft},
   year = {2022},
   version = {0.2.16},
-  url = {https://cran.r-project.org/web/packages/miniCRAN/index.html}
+  url = {https://cran.r-project.org/web/packages/miniCRAN/index.html},
+  doi = {10.32614/cran.package.minicran}
 }
 
 @misc{cia2021internetusers,
@@ -75,13 +77,13 @@
   url = {https://pypi.org/project/requests/}
 }
 
-@book{rossum2009py, 
- title = {Python 3 Reference Manual}, 
- author = {Van Rossum, Guido and Drake, Fred L.}, 
- year = {2009}, 
- isbn = {1441412697}, 
- publisher = {CreateSpace}, 
- address = {Scotts Valley, CA} 
+@book{rossum2009py,
+ title = {Python 3 Reference Manual},
+ author = {Van Rossum, Guido and Drake, Fred L.},
+ year = {2009},
+ isbn = {1441412697},
+ publisher = {CreateSpace},
+ address = {Scotts Valley, CA}
 }
 
 @misc{fsf2010wget,
@@ -90,7 +92,7 @@
   note = {Original author: Hrvoje Nikšić},
   url = {http://www.gnu.org/software/wget/},
   version = {1.24.5},
-  year = {2024}, 
+  year = {2024},
 }
 
 @inproceedings{soton403913,
@@ -102,7 +104,8 @@
             year = {2016},
            pages = {87--90},
              url = {https://eprints.soton.ac.uk/403913/},
-        abstract = {It is increasingly necessary for researchers in all fields to write computer code, and in order to reproduce research results, it is important that this code is published. We present Jupyter notebooks, a document format for publishing code, results and explanations in a form that is both readable and executable. We discuss various tools and use cases for notebook documents.}
+        abstract = {It is increasingly necessary for researchers in all fields to write computer code, and in order to reproduce research results, it is important that this code is published. We present Jupyter notebooks, a document format for publishing code, results and explanations in a form that is both readable and executable. We discuss various tools and use cases for notebook documents.},
+        doi = {10.3233/978-1-61499-649-1-87}
 }
 
 @Manual{rstudio2024,
@@ -120,6 +123,7 @@
   year = {2024},
   version  = {2.5.0},
   url = {https://cran.r-project.org/web/packages/remotes/index.html},
+  doi = {10.32614/cran.package.remotes}
   }
 
 @Manual{r2024,
@@ -129,9 +133,5 @@
     address = {Vienna, Austria},
     year = {2024},
     url = {https://www.R-project.org/},
+    doi = {10.32614/r.manuals}
   }
-  
-
-
-
-
