### 2026-02-16 1c85651
As pointed out in #151 it would be really difficult to understand the package just from our CLI help. This PR expands the help text to try to improve this.
