### 2026-05-19 fea0bbc
Remove software documentation
This documentation is not supposed to be included in JOSE papers and
this also helps reduce the length of the paper.
