### 2026-05-19 7792de7
Tighten Software Design section of paper
