### 2026-05-21 ffb73f4
Improve installation docs
- More detailed information on installation
- Discuss R dependency and $ installation
- Just link to Installation page from Getting Started instead of
  repeating ourselves

Closes #157
