### 2026-05-22 fd1ec87
Indicate in paper summary that servers are part of larger ecosystem
