### 2026-05-22 3f030ec
Update size of installer downloads
