### 2026-05-23 8c4e98c
Fix invalid escape warning in R version regex
