+++ docs/index.rst
@@ -27,6 +27,7 @@ tools and instructional materials for teaching and practicing data science in in
 
    installation
    getting-started
+   using-downloads
    statement-of-need
    contributing
    code-of-conduct
