### 2026-05-27 23bc1e6
Add page discussing what to install and associated data sizes
