### 2026-05-29 4f71cd6
Enable GitHub PAT for remotes install of miniCRAN
This change can be reverted once there's a new release of miniCRAN that
we can install normally
