+++ .github/workflows/package-check.yml
@@ -8,6 +8,9 @@ on:
 
 name: Build Check
 
+env:
+  GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }} # Can remove once installing miniCRAN from CRAN again
+
 jobs:
   build-check:
     runs-on: ${{ matrix.config.os }}
