### 2026-05-29 b1298e2
Add progress bars to provide clearer progress info to users
Closes #152
